create oneTrustAssessmentToCsv helper
