+++ src/oneTrust/helpers/writeOneTrustAssessment.ts
@@ -2,12 +2,9 @@ import { logger } from '../../logger';
 import colors from 'colors';
 import { OneTrustFileFormat } from '../../enums';
 import fs from 'fs';
-import { flattenOneTrustAssessment } from './flattenOneTrustAssessment';
-import { DEFAULT_ONE_TRUST_ASSESSMENT_CSV_HEADER } from '../constants';
-import { decodeCodec } from '@transcend-io/type-utils';
-import { OneTrustAssessmentCsvRecord } from '@transcend-io/privacy-types';
 import { OneTrustEnrichedAssessment } from '../codecs';
 import { oneTrustAssessmentToJson } from './oneTrustAssessmentToJson';
+import { oneTrustAssessmentToCsv } from './oneTrustAssessmentToCsv';
 
 /**
  * Write the assessment to disk at the specified file path.
@@ -41,46 +38,16 @@ export const writeOneTrustAssessment = ({
     ),
   );
 
-  // For json format
   if (fileFormat === OneTrustFileFormat.Json) {
-    const jsonEntry = oneTrustAssessmentToJson({
-      assessment,
-      index,
-      total,
-    });
-    fs.appendFileSync(file, jsonEntry);
-  } else if (fileFormat === OneTrustFileFormat.Csv) {
-    const csvRows = [];
-
-    // write csv header at the beginning of the file
-    if (index === 0) {
-      csvRows.push(DEFAULT_ONE_TRUST_ASSESSMENT_CSV_HEADER.join(','));
-    }
-
-    // flatten the assessment object so it does not have nested properties
-    const flatAssessment = flattenOneTrustAssessment(assessment);
-
-    // comment
-    const flatAssessmentFull = Object.fromEntries(
-      DEFAULT_ONE_TRUST_ASSESSMENT_CSV_HEADER.map((header) => {
-        const value = flatAssessment[header] ?? '';
-        const escapedValue =
-          typeof value === 'string' &&
-          (value.includes(',') || value.includes('"'))
-            ? `"${value.replace(/"/g, '""')}"`
-            : value;
-        return [header, escapedValue];
+    fs.appendFileSync(
+      file,
+      oneTrustAssessmentToJson({
+        assessment,
+        index,
+        total,
       }),
     );
-
-    // ensure the record has the expected type!
-    decodeCodec(OneTrustAssessmentCsvRecord, flatAssessmentFull);
-
-    // transform the flat assessment to have all CSV keys in the expected order
-    const assessmentRow = Object.values(flatAssessmentFull);
-
-    // append the rows to the file
-    csvRows.push(`${assessmentRow.join(',')}\n`);
-    fs.appendFileSync('./oneTrust.csv', csvRows.join('\n'));
+  } else if (fileFormat === OneTrustFileFormat.Csv) {
+    fs.appendFileSync(file, oneTrustAssessmentToCsv({ assessment, index }));
   }
 };
